Add cloaking and de-cloaking.
It's still not clear when and how I will use this, exactly.. but at
least it is working now.  Next up is looking into the how and whys of
chunking.

Also moves the "hex" option into the test code.  I don't think I'll need
it for running code.

Thanks to @quartzjer for help and advice on this.